Update messages.json (es)
I speak spanish and decided to help a little bit to improve the translation. I mainly change some options to mirror the Windows' options names in spanish and also some errors in the wors "save" ("guardar" instead of "ahorro") and "right" ("derecha" instead of "correcto"). These two errors always show up because there is no context when is translated.